Fix up init6 to only boot once

+++ scratch/recipes/init6.rb
@@ -1,2 +1,7 @@
-log "Issueing an init6 to simulate an early reboot"
-execute "init 6"
+if ::File.exists?("/var/lib/booted")
+  log "Skipping reboot, /var/lib/booted file indicates we already booted"
+else
+  log "Issueing an init6 to simulate an early reboot"
+  ::File.open("/var/lib/booted","w") {|f| f.puts(Time.now)}
+  execute "init 6"
+end
